fixed descriptions and added class of error
