+++ app.py
@@ -2,7 +2,7 @@ import json
 from flask import Flask, request, jsonify
 from flask.views import MethodView
 from werkzeug.exceptions import HTTPException
-from HandledError import HandledError
+from error_handling import HandledError, ValidationError
 
 app = Flask(__name__)
 pattern = {'id': int, 'name': str, 'address': str, "work_time": str, "phone_number": str}
@@ -14,24 +14,20 @@ restaurants = data['restaurants']
 
 def validate_post(content):
     if content is None:
-        raise HandledError(status_code=400, description="Required data not available", name="Bad Request")
+        raise ValidationError(description="Required data not available(No data)")
     if content.keys() != pattern.keys():
-        raise HandledError(status_code=400, description="Sent data not correct", name="Bad Request")
-
-    if any([type(content['id']) != pattern['id'],
-            type(content['name']) != pattern['name'],
-            type(content['address']) != pattern['address'],
-            type(content['work_time']) != pattern['work_time'],
-            type(content['phone_number']) != pattern['phone_number']]):
-        raise HandledError(status_code=400, description="Type of data not correct", name="Bad Request")
+        raise ValidationError(description="Sent data not correct, doesn't match format")
+    for key, value in pattern.items():
+        if type(content[key]) != value:
+            raise ValidationError(description="Type of sent data not correct")
 
 
 def validate_put(content):
     if content is None:
-        raise HandledError(status_code=400, description="Required data not available", name="Bad Request")
+        raise ValidationError(description="Required data not available(No data)")
     for key in content.keys():
         if key not in pattern.keys():
-            raise HandledError(status_code=400, description="Required data not correct.", name="Bad Request")
+            raise ValidationError(description="Required data not correct, doesn't match format")
 
 
 class RestaurantEndpoint(MethodView):
@@ -50,7 +46,7 @@ class RestaurantItemEndpoint(MethodView):
         for restaurant in restaurants:
             if restaurant["id"] == restaurant_id:
                 return jsonify(restaurant), 200
-        raise(HandledError(status_code=404, description="Page doesn't exist", name="Not Found"))
+        raise(HandledError(status_code=404, description="Restaurant with such ID doesn't exist", name="Not Found"))
 
     def put(self, restaurant_id):
         content = request.json
@@ -60,7 +56,9 @@ class RestaurantItemEndpoint(MethodView):
                 for key, value in content.items():
                     restaurant[key] = value
                 return jsonify(restaurant), 200
-        raise (HandledError(status_code=404, description="Page doesn't exist", name="Not Found"))
+        raise (HandledError(status_code=404,
+                            description="No restaurant to update. Restaurant with such ID doesn't exist",
+                            name="Not Found"))
 
 
 app.add_url_rule("/api/restaurants", view_func=RestaurantEndpoint.as_view("restaurant_api"))
@@ -82,6 +80,7 @@ def handle_standard_exception(ex):
         "name": ex.name,
         "description": ex.description,
     }), ex.code
+
 
 if __name__ == '__main__':
     app.run()
